requirements.txt
playwright==1.48.0
pyinstaller==6.11.0
openpyxl==3.1.5
python-dotenv==1.0.1
